updates for 0.10.2 and integration of a few patches, specifically
 a slightly altered version of moshe zadka's client timeouts

 the remainder of chris hadgis's patch for tkfaucet

 a fix to tkinternet to make sure the timeout's always an int

 documentation on what a plugin is


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@414 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/internet/tkinternet.py
@@ -45,7 +45,7 @@ class Simulator:
         timeout = main.runUntilCurrent()
         if timeout is not None:
             _sim = Simulator()
-            _root.after(timeout * 1010, _sim.simulate)
+            _root.after(int(timeout * 1010), _sim.simulate)
 
     def cancel(self):
         self.cancelled = 1
